Added token limit

+++ main.py
@@ -5,7 +5,7 @@ import tempfile
 from langchain.chat_models import ChatOpenAI
 
 
-from utils import doc_loader, auto_summary_builder, check_key_validity, summary_prompt_creator, check_gpt_4
+from utils import doc_loader, auto_summary_builder, check_key_validity, summary_prompt_creator, check_gpt_4, token_limit
 from my_prompts import map_prompt, combine_prompt
 from file_conversions import pdf_to_text
 
@@ -19,7 +19,13 @@ st.sidebar.markdown('# Made by: [Ethan](https://github.com/e-johnstonn)')
 st.sidebar.markdown('# Git link: [Docsummarizer](https://github.com/e-johnstonn/docsummarizer)')
 
 
-if st.button("Summarize"):
+
+if 'summarize_button_clicked' not in st.session_state:
+    st.session_state.summarize_button_clicked = False
+
+
+if st.button('Summarize', disabled=st.session_state.summarize_button_clicked):
+    st.session_state.summarize_button_clicked = True
     valid = check_key_validity(api_key)
     valid_gpt_4 = True
     if use_gpt_4:
@@ -37,16 +43,22 @@ if st.button("Summarize"):
                 llm = ChatOpenAI(openai_api_key=api_key, temperature=0, max_tokens=500, model_name='gpt-3.5-turbo')
             else:
                 llm = ChatOpenAI(openai_api_key=api_key, temperature=0, max_tokens=250, model_name='gpt-3.5-turbo')
+
             initial_chain = summary_prompt_creator(map_prompt, 'text', llm)
             final_prompt_list = summary_prompt_creator(combine_prompt, 'text', llm)
             doc = doc_loader(temp_file_path)
-            summary = auto_summary_builder(doc, 10, initial_chain, final_prompt_list, api_key, use_gpt_4)
-            st.markdown(summary, unsafe_allow_html=True)
-            os.unlink(temp_file_path)
+            limit_check = token_limit(doc, 8000)
+            if limit_check:
+                summary = auto_summary_builder(doc, 10, initial_chain, final_prompt_list, api_key, use_gpt_4)
+                st.markdown(summary, unsafe_allow_html=True)
+                os.unlink(temp_file_path)
+            else:
+                st.warning('File too big!')
     elif uploaded_file is None:
         st.warning("Please upload a file.")
     elif valid is True:
         st.warning(check_gpt_4(api_key))
     else:
         st.warning(check_key_validity(api_key))
+    st.session_state.summarize_button_clicked = False
